# openapi/zeus.v1.yaml
openapi: 3.1.0
info:
  title: Zeus API
  description: Enterprise Harness Agent control plane and durable execution API.
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
paths:
...
  /api/v1/workspaces/{workspace_id}/model-profiles:
    get:
      tags:
      - models-and-tools
      operationId: list_model_profiles
      parameters:
      - name: workspace_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfilePageResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
    post:
      tags:
      - models-and-tools
      operationId: create_model_profile
      parameters:
      - name: workspace_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      requestBody:
        description: JSON request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelProfileRequest'
        required: true
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
  /api/v1/workspaces/{workspace_id}/model-profiles/{model_profile_id}:
    get:
      tags:
      - models-and-tools
      operationId: get_model_profile
      parameters:
      - name: workspace_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      - name: model_profile_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
    patch:
      tags:
      - models-and-tools
      operationId: update_model_profile
      parameters:
      - name: workspace_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      - name: model_profile_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      requestBody:
        description: JSON request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelProfileRequest'
        required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
  /api/v1/workspaces/{workspace_id}/model-profiles/{model_profile_id}/archive:
    post:
      tags:
      - models-and-tools
      operationId: archive_model_profile
      parameters:
      - name: workspace_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      - name: model_profile_id
        in: path
        description: Path identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileResponse'
        default:
          description: Problem Details error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - sessionCookie: []
      - serviceAccountBearer: []
...
components:
  schemas:
...
    CreateModelProfileRequest:
      type: object
      required:
      - connection_id
      - name
      - base_url
      - model
      properties:
        base_url:
          type: string
        configuration: {}
        connection_id:
          type: string
          format: uuid
        model:
          type: string
        name:
          type: string
        provider_kind:
          type: string
...
    ModelProfilePageResponse:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ModelProfileResponse'
        next_cursor:
          type:
          - string
          - 'null'
    ModelProfileResponse:
      type: object
      required:
      - id
      - organization_id
      - workspace_id
      - connection_id
      - name
      - provider_kind
      - base_url
      - model
      - configuration
      - revision
      - created_at
      - updated_at
      properties:
        archived_at:
          type:
          - string
          - 'null'
          format: date-time
        base_url:
          type: string
        configuration: {}
        connection_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        model:
          type: string
        name:
          type: string
        organization_id:
          type: string
          format: uuid
        provider_kind:
          type: string
        revision:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time
        workspace_id:
          type: string
          format: uuid
...
    ProblemDetails:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
          format: uuid
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          example: https://zeus.example.com/problems/validation_failed
...
    UpdateModelProfileRequest:
      type: object
      properties:
        archived:
          type:
          - boolean
          - 'null'
        base_url:
          type:
          - string
          - 'null'
        configuration: {}
        connection_id:
          type:
          - string
          - 'null'
          format: uuid
        model:
          type:
          - string
          - 'null'
        name:
          type:
          - string
          - 'null'
        provider_kind:
          type:
          - string
          - 'null'
...
  securitySchemes:
    serviceAccountBearer:
      type: http
      scheme: bearer
      bearerFormat: ServiceAccount
      description: Bearer token for a Zeus service account
    sessionCookie:
      type: apiKey
      in: cookie
      name: zeus_session
      description: HttpOnly session cookie
...